Euler Tour
"It's working It's working!!"
-Anakin Skywalker

+++ TSPCS325/vertex.cpp
@@ -3,6 +3,6 @@
 
 int getDistance(struct vertex *v1, struct vertex *v2) {
 	return round(
-		(long double) sqrt((pow(v1->xCoord - v2->xCoord, 2)) + (pow(v1->yCoord - v2->yCoord, 2)))
+		sqrt((pow(v1->xCoord - v2->xCoord, 2)) + (pow(v1->yCoord - v2->yCoord, 2)))
 	);
 }
